Break dag down into some libs
