+++ tools/dag/index.ts
@@ -1,51 +1,14 @@
-import glob from "glob";
-import { readFile, writeFile } from "fs/promises";
+import { writeFile } from "fs/promises";
 
 import { getChangedFiles } from "./lib/git";
-
-type ModuleName = string;
-
-interface Deps {
-  dependsOn: ModuleName[];
-  name?: string;
-}
-
-interface DepsMap {
-  [path: string]: Deps;
-}
-
-const loadDepsJson = async (
-  path: string
-): Promise<{ module: ModuleName; path: string; deps: Deps }> => {
-  const file = await readFile(path, { encoding: "utf-8" });
-
-  try {
-    const deps = JSON.parse(file) as Deps;
-    return { module: path.replace("/deps.json", ""), deps, path };
-  } catch (error) {
-    console.error(`Failed to parse ${path}`, error);
-    throw error;
-  }
-};
-
-const getDepsFiles = (): Promise<string[]> =>
-  new Promise((resolve, reject) => {
-    glob("**/deps.json", {}, (error, matches) => {
-      if (error) {
-        console.error(error);
-        return reject(error);
-      }
-
-      console.log("Found some deps!", matches);
-      console.log();
-
-      resolve(matches);
-    });
-  });
+import { getDepsMap } from "./lib/deps";
+import { getModulesToBuild } from "./lib/modules";
+import type { DepsMap } from "./types/dag";
 
 (async () => {
   process.chdir("../../");
 
+  // Step one - Pull all the file changes on this branch from Git
   let changedFiles: string[] = [];
 
   try {
@@ -58,64 +21,32 @@ const getDepsFiles = (): Promise<string[]> =>
     return process.exit(1);
   }
 
+  // Step two - Filter the changes to code in apps or libs
+  // TODO: The paths to check can be moved into config
   const filteredChangedFiles = changedFiles.filter(
-    (path) => path.startsWith("apps/") || path.startsWith("lib/")
+    (path) => path.startsWith("apps/") || path.startsWith("libs/")
   );
 
   console.log("Changes affecting build are", filteredChangedFiles);
   console.log();
 
+  // Step three - Load all the deps files in the repo so we have a full list of modules and their deps
   let depsMap: DepsMap = {};
-
   try {
-    const files = await getDepsFiles();
-    const depMaps = await Promise.all(files.map(loadDepsJson));
-
-    depsMap = depMaps.reduce<DepsMap>((acc, map) => {
-      return {
-        ...acc,
-        [map.module]: map.deps,
-      };
-    }, {});
+    depsMap = await getDepsMap();
   } catch (error) {
-    console.error(error);
     return process.exit(1);
   }
 
   console.log("Deps map is", depsMap);
   console.log();
 
-  const toBuild = filteredChangedFiles.reduce<{ [moduleName: ModuleName]: boolean }>((acc, path) => {
-    const moduleMatch = /(?:apps|libs)\/\w+/.exec(path);
-    if (!moduleMatch || moduleMatch.length > 1) {
-      console.warn('Bad module match detected. Got:', moduleMatch);
-      return acc;
-    }
-
-    const moduleName = moduleMatch[0];
-
-    const isApp = moduleName.startsWith('apps/');
-    if (isApp && acc[moduleMatch[0]]) {
-      console.info('App', moduleName, 'is already marked for build');
-      return acc;
-    }
-
-    if (!isApp) {
-      // TODO: Need to do a search on the DAG to work out which apps to build
-      console.info("Module -> app resolution not yet implemented");
-      return acc;
-    }
-
-    console.log("New app to build detected:", moduleName);
-
-    return {
-      ...acc,
-      [moduleName]: true
-    };
-  }, {});
+  // Step four - Work out what we need to build
+  const toBuild = getModulesToBuild(filteredChangedFiles);
 
   console.log();
   console.log('Modules to build are', toBuild);
 
+  // Step five - Write the build list to a file ready to be passed to docker buildx bake
   await writeFile('./toBuild.json', JSON.stringify(toBuild, null, 2));
 })();
