+++ tools/dag/types/dag.d.ts
@@ -1,0 +1,10 @@
+export type ModuleName = string;
+
+export interface Deps {
+  dependsOn: ModuleName[];
+  name?: string;
+}
+
+export interface DepsMap {
+  [path: string]: Deps;
+}
